+++ mesa/__init__.py
@@ -7,6 +7,11 @@ Core Objects: Model, and Agent.
 import datetime as dt
 import random
 
+__title__ = 'mesa'
+__version__ = '0.6.5'
+__license__ = 'Apache 2.0'
+__copyright__ = 'Copyright 2015 Project Mesa Team'
+
 
 class Model(object):
     '''